#003 renamed authBannerContainer
